fix: debian #Build should added steps

+++ cuepkg/debian/build.cue
@@ -18,16 +18,17 @@ import (
 	platform?: string
 	mirror?:   string
 	packages: [pkgName=string]: string | *""
+	steps: [...docker.#Step]
 
 	_build: docker.#Build & {
-		steps: [
+		"steps": [
 			docker.#Pull & {
 				"source": source
 				if platform != _|_ {
 					"platform": platform
 				}
 			},
-			if mirror != _|_ {
+			if mirror != _|_ && len(packages) > 0 {
 				docker.#Run & {
 					_script: [
 							if mirror != "" {"""
@@ -42,15 +43,20 @@ import (
 					}
 				}
 			},
-			docker.#Run & {
-				command: {
-					name: "sh"
-					flags: "-c": strings.Join([
-							"apt-get update -y",
-							for _pkgName, _version in packages {"apt-get install -y -f \(_pkgName)\(_version)"},
-							"rm -rf /var/lib/apt/lists/*",
-					], "\n")
+			if len(packages) > 0 {
+				docker.#Run & {
+					command: {
+						name: "sh"
+						flags: "-c": strings.Join([
+								"apt-get update -y",
+								for _pkgName, _version in packages {"apt-get install -y -f \(_pkgName)\(_version)"},
+								"rm -rf /var/lib/apt/lists/*",
+						], "\n")
+					}
 				}
+			},
+			for _, step in steps {
+				step
 			},
 		]
 	}
